Move classes to seperate files and add xmldoc
