+++ Terminal.Gui/Drivers/V2/ApplicationV2.cs
@@ -271,73 +271,3 @@ public class ApplicationV2 : ApplicationImpl
         Application.Top?.SetNeedsLayout ();
     }
 }
-
-public class NetComponentFactory : ComponentFactory<ConsoleKeyInfo>
-{
-    public override IConsoleInput<ConsoleKeyInfo> CreateInput ()
-    {
-        return new NetInput ();
-    }
-
-    /// <inheritdoc />
-    public override IConsoleOutput CreateOutput ()
-    {
-        return new NetOutput ();
-    }
-
-    /// <inheritdoc />
-    public override IInputProcessor CreateInputProcessor (ConcurrentQueue<ConsoleKeyInfo> inputBuffer)
-    {
-        return new NetInputProcessor (inputBuffer);
-    }
-}
-
-public class WindowsComponentFactory : ComponentFactory<WindowsConsole.InputRecord>
-{
-    /// <inheritdoc />
-    public override IConsoleInput<WindowsConsole.InputRecord> CreateInput ()
-    {
-        return new WindowsInput ();
-    }
-
-    /// <inheritdoc />
-    public override IInputProcessor CreateInputProcessor (ConcurrentQueue<WindowsConsole.InputRecord> inputBuffer)
-    {
-        return new WindowsInputProcessor (inputBuffer);
-    }
-
-    /// <inheritdoc />
-    public override IConsoleOutput CreateOutput ()
-    {
-        return new WindowsOutput ();
-    }
-}
-
-public abstract class ComponentFactory<T> : IComponentFactory<T>
-{
-    /// <inheritdoc />
-    public abstract IConsoleInput<T> CreateInput ();
-
-    /// <inheritdoc />
-    public abstract IInputProcessor CreateInputProcessor (ConcurrentQueue<T> inputBuffer);
-
-    /// <inheritdoc />
-    public virtual IWindowSizeMonitor CreateWindowSizeMonitor (IConsoleOutput consoleOutput, IOutputBuffer outputBuffer)
-    {
-        return new WindowSizeMonitor (consoleOutput, outputBuffer);
-    }
-
-    /// <inheritdoc />
-    public abstract IConsoleOutput CreateOutput ();
-}
-public interface IComponentFactory<T> : IComponentFactory
-{
-    IConsoleInput<T> CreateInput ();
-    IInputProcessor CreateInputProcessor (ConcurrentQueue<T> inputBuffer);
-    IWindowSizeMonitor CreateWindowSizeMonitor (IConsoleOutput consoleOutput, IOutputBuffer outputBuffer);
-}
-
-public interface IComponentFactory
-{
-    IConsoleOutput CreateOutput ();
-}
